chore: relocate play/pause button
