--- src/components/Button/button.module.scss
@@ -1,26 +1,0 @@
-@use '@/utils/variables';
-
-.button {
-  border: none;
-  background-color: inherit;
-  padding: 0 !important;
-  width: fit-content;
-  height: fit-content;
-
-  svg {
-    width: 1rem;
-    height: 1rem;
-    padding: 0.2rem;
-    border-radius: 2px;
-    background-color: variables.$white;
-    color: variables.$raisin-black;
-    box-shadow: variables.$button-shadow;
-  }
-}
-.button:hover {
-  svg {
-    background-color: variables.$pale-white-grey;
-  }
-
-  cursor: pointer;
-}
