+++ src/components/_shared/Button/PauseButton.tsx
@@ -1,16 +1,15 @@
 import { pauseTrack } from '@/api/spotify';
 import PauseIcon from '@/assets/pauseIcon.svg';
-import { useDebounce } from '@/hooks/useDebounce';
 import classNames from 'classnames/bind';
-import Style from '../button.module.scss';
+import Style from './button.module.scss';
 
 const style = classNames.bind(Style);
 
 const PauseButton = () => {
-  const onPauseDebounceHandler = useDebounce(() => pauseTrack({}));
+  const handlePause = () => pauseTrack({});
 
   return (
-    <button className={style('button')} onClick={onPauseDebounceHandler}>
+    <button className={style('button')} onClick={handlePause}>
       <PauseIcon />
     </button>
   );
