--- src/components/_shared/Button/PlayPauseButton/PlayButton.tsx
@@ -1,36 +1,0 @@
-import { playTrack } from '@/api/spotify';
-import PlayIcon from '@/assets/playIcon.svg';
-import { useDebounce } from '@/hooks/useDebounce';
-import { MetaInfo } from '@/models/MetaInfo';
-import { CurrentlyPlayingTrack } from '@/models/track';
-import classNames from 'classnames/bind';
-import { HtmlHTMLAttributes } from 'react';
-import Style from '../button.module.scss';
-
-const style = classNames.bind(Style);
-
-export interface PlayButtonProps extends HtmlHTMLAttributes<HTMLButtonElement> {
-  context: MetaInfo['uri'];
-  uri?: CurrentlyPlayingTrack['item']['uri'];
-  position_ms: CurrentlyPlayingTrack['progress_ms'];
-}
-
-const PlayButton = ({ context, uri, position_ms }: PlayButtonProps) => {
-  const onPlayDebounceHandler = useDebounce(() => {
-    playTrack({
-      offset: { uri },
-      position_ms,
-      context_uris: context,
-    });
-  });
-
-  return (
-    <button
-      className={style('play-button', 'button')}
-      onClick={onPlayDebounceHandler}
-    >
-      <PlayIcon />
-    </button>
-  );
-};
-export default PlayButton;
